Extract IO to new class

+++ src/openrct2/localisation/Localisation.Date.cpp
@@ -14,7 +14,7 @@
 #include <algorithm>
 #include <time.h>
 
-uint16_t gDateMonthTicks;
+uint32_t gDateMonthTicks;
 int32_t gDateMonthsElapsed;
 
 // rct2: 0x00993988
